examples/image_in/requirements.txt
opencv-python>=4.5.0
boto3>=1.26.0
google-cloud-storage>=2.0.0
numpy>=1.21.0
fastapi>=0.68.0
uvicorn>=0.15.0